requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.2
